refactor(sidebar): extract DnD hook, context menu, section mutations; show drop targets during drag
- Extract useWorkspaceDnD hook from WorkspaceListItem (DnD setup, reorder, move-to-section on drop)
- Extract WorkspaceContextMenu component (context menu items, hover card wrapping)
- Extract useSectionMutations hook from WorkspaceSection (toggle/rename/delete/setColor)
- Add onSettled invalidation to reorder callbacks to restore server state on failure
- Move WORKSPACE_DND_TYPE to constants.ts
- Show drop zone indicators when any compatible item is dragging (not just on hover)
- useSectionDropZone now returns isDropTarget (any valid drag active) alongside isDragOver
