+++ apps/desktop/src/renderer/screens/main/components/WorkspaceSidebar/WorkspaceListItem/index.ts
@@ -1,7 +1,5 @@
 export type { DragItem } from "../types";
+export { WORKSPACE_DND_TYPE } from "./constants";
 export { WorkspaceDiffStats } from "./WorkspaceDiffStats";
-export {
-	WORKSPACE_DND_TYPE,
-	WorkspaceListItem,
-} from "./WorkspaceListItem";
+export { WorkspaceListItem } from "./WorkspaceListItem";
 export { WorkspaceStatusBadge } from "./WorkspaceStatusBadge";
