+++ apps/desktop/src/renderer/screens/main/components/WorkspaceSidebar/WorkspaceListItem/useWorkspaceDnD.ts
@@ -1,0 +1,162 @@
+import { toast } from "@superset/ui/sonner";
+import { useCallback } from "react";
+import { useDrag, useDrop } from "react-dnd";
+import { electronTrpc } from "renderer/lib/electron-trpc";
+import {
+	useMoveWorkspacesToSection,
+	useMoveWorkspaceToSection,
+	useReorderWorkspaces,
+	useReorderWorkspacesInSection,
+} from "renderer/react-query/workspaces";
+import { invalidateWorkspaceQueries } from "renderer/react-query/workspaces/invalidateWorkspaceQueries";
+import { useActiveDragItemStore } from "renderer/stores/active-drag-item";
+import { useWorkspaceSelectionStore } from "renderer/stores/workspace-selection";
+import type { DragItem } from "../types";
+import { WORKSPACE_DND_TYPE } from "./constants";
+
+interface UseWorkspaceDnDOptions {
+	id: string;
+	projectId: string;
+	sectionId: string | null;
+	index: number;
+}
+
+export function useWorkspaceDnD({
+	id,
+	projectId,
+	sectionId,
+	index,
+}: UseWorkspaceDnDOptions) {
+	const utils = electronTrpc.useUtils();
+	const reorderWorkspaces = useReorderWorkspaces();
+	const reorderWorkspacesInSection = useReorderWorkspacesInSection();
+	const moveToSection = useMoveWorkspaceToSection();
+	const bulkMoveToSection = useMoveWorkspacesToSection();
+	const selectionStore = useWorkspaceSelectionStore;
+
+	const handleReorder = useCallback(
+		(item: DragItem) => {
+			if (item.originalIndex === item.index) return;
+			const callbacks = {
+				onError: (error: { message: string }) => {
+					void invalidateWorkspaceQueries(utils);
+					toast.error(`Failed to reorder workspace: ${error.message}`);
+				},
+				onSettled: () => invalidateWorkspaceQueries(utils),
+			};
+			if (item.sectionId !== null) {
+				reorderWorkspacesInSection.mutate(
+					{
+						sectionId: item.sectionId,
+						fromIndex: item.originalIndex,
+						toIndex: item.index,
+					},
+					callbacks,
+				);
+			} else {
+				reorderWorkspaces.mutate(
+					{
+						projectId: item.projectId,
+						fromIndex: item.originalIndex,
+						toIndex: item.index,
+					},
+					callbacks,
+				);
+			}
+		},
+		[reorderWorkspaces, reorderWorkspacesInSection, utils],
+	);
+
+	const [{ isDragging }, drag] = useDrag(
+		() => ({
+			type: WORKSPACE_DND_TYPE,
+			item: () => {
+				const selection = selectionStore.getState();
+				const isPartOfSelection = selection.selectedIds.has(id);
+				if (!isPartOfSelection) {
+					selection.clearSelection();
+				}
+				const selectedIds =
+					isPartOfSelection && selection.selectedIds.size > 1
+						? [...selection.selectedIds]
+						: undefined;
+				const dragItem: DragItem = {
+					id,
+					projectId,
+					sectionId,
+					index,
+					originalIndex: index,
+					selectedIds,
+				};
+				useActiveDragItemStore.getState().setActiveDragItem(dragItem);
+				return dragItem;
+			},
+			end: (item, monitor) => {
+				useActiveDragItemStore.getState().clearActiveDragItem();
+				selectionStore.getState().clearSelection();
+				if (!item) return;
+				if (item.handled || monitor.didDrop()) return;
+				handleReorder(item);
+			},
+			collect: (monitor) => ({ isDragging: monitor.isDragging() }),
+		}),
+		[id, projectId, sectionId, index, handleReorder],
+	);
+
+	const [, drop] = useDrop({
+		accept: WORKSPACE_DND_TYPE,
+		hover: (item: DragItem) => {
+			if (item.selectedIds && item.selectedIds.length > 1) return;
+			if (
+				item.projectId !== projectId ||
+				item.sectionId !== sectionId ||
+				item.index === index
+			)
+				return;
+			utils.workspaces.getAllGrouped.setData(undefined, (oldData) => {
+				if (!oldData) return oldData;
+				return oldData.map((group) => {
+					if (group.project.id !== projectId) return group;
+					if (sectionId === null) {
+						const workspaces = [...group.workspaces];
+						const [moved] = workspaces.splice(item.index, 1);
+						workspaces.splice(index, 0, moved);
+						return { ...group, workspaces };
+					}
+					const sections = group.sections.map((section) => {
+						if (section.id !== sectionId) return section;
+						const workspaces = [...section.workspaces];
+						const [moved] = workspaces.splice(item.index, 1);
+						workspaces.splice(index, 0, moved);
+						return { ...section, workspaces };
+					});
+					return { ...group, sections };
+				});
+			});
+			item.index = index;
+		},
+		drop: (item: DragItem) => {
+			if (item.projectId !== projectId) return;
+			if (item.sectionId === sectionId) {
+				handleReorder(item);
+				if (item.originalIndex !== item.index) return { reordered: true };
+			} else if (!item.handled) {
+				if (item.selectedIds && item.selectedIds.length > 1) {
+					bulkMoveToSection.mutate({
+						workspaceIds: item.selectedIds,
+						sectionId,
+					});
+				} else {
+					moveToSection.mutate({
+						workspaceId: item.id,
+						sectionId,
+					});
+				}
+				item.handled = true;
+				return { moved: true };
+			}
+		},
+	});
+
+	return { isDragging, drag, drop };
+}
